% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decisionTreeR.R
\name{decisionTreeR}
\alias{decisionTreeR}
\title{Decision Tree Classification Algorithm (R Version)}
\usage{
decisionTreeR(data, depth = 1, max_depth = 5)
}
\arguments{
\item{data}{Training dataset, including features and labels}

\item{depth}{Current tree depth, default is 1}

\item{max_depth}{Maximum tree depth, default is 5}
}
\value{
Decision tree model
}
\description{
Implementation of a decision tree classification algorithm in R. The tree depth does not exceed 5 and uses the information gain method for attribute splitting.
}
\examples{
\dontrun{
data(filmtrain)
tree <- decisionTreeR(filmtrain)
}
}
