% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{informationGainCpp}
\alias{informationGainCpp}
\title{信息增益计算（Rcpp 版本）}
\usage{
informationGainCpp(data, feature)
}
\arguments{
\item{data}{数据框，包含特征和标签}

\item{feature}{要计算的信息增益的特征名称}
}
\value{
信息增益值
}
\description{
使用 Rcpp 实现的信息增益计算函数
}
\examples{
\dontrun{
data(filmtrain)
gain <- informationGainCpp(filmtrain, "V1")
}
}
