% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kmeansCpp}
\alias{kmeansCpp}
\title{K-Means 聚类算法（Rcpp 版本）}
\usage{
kmeansCpp(X, k, max_iter = 100L)
}
\arguments{
\item{X}{数据矩阵，每行是一个样本}

\item{k}{聚类的类别数}

\item{max_iter}{最大迭代次数，默认为 100}
}
\value{
一个列表，包含聚类结果和聚类中心
}
\description{
使用 Rcpp 实现的 K-Means 聚类算法
}
\examples{
\dontrun{
data(ori_data_x)
result <- kmeansCpp(as.matrix(ori_data_x), k = 3)
}
}
