% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{naiveBayesCpp}
\alias{naiveBayesCpp}
\title{朴素贝叶斯分类算法（Rcpp 版本）}
\usage{
naiveBayesCpp(data)
}
\arguments{
\item{data}{训练数据集，包括特征和标签}
}
\value{
训练好的朴素贝叶斯模型
}
\description{
使用 Rcpp 实现的朴素贝叶斯分类算法
}
\examples{
\dontrun{
data(filmtrain)
model <- naiveBayesCpp(as.matrix(filmtrain))
}
}
