% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/naiveBayesR.R
\name{naiveBayesR}
\alias{naiveBayesR}
\title{朴素贝叶斯分类算法（R版本）}
\usage{
naiveBayesR(data)
}
\arguments{
\item{data}{训练数据集，包括特征和标签}
}
\value{
训练好的朴素贝叶斯模型
}
\description{
使用 R 语言实现的朴素贝叶斯分类算法
}
\examples{
\dontrun{
data(filmtrain)
model <- naiveBayesR(filmtrain)
}
}
