% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_docs.R
\docType{data}
\name{ori_data}
\alias{ori_data}
\title{Original Data for Analysis}
\format{
A data frame with 2000 rows and 16 variables:
\describe{
  \item{var1}{Numeric variable representing ...}
  \item{var2}{Numeric variable representing ...}
  \item{var3}{Numeric variable representing ...}
  \item{var4}{Numeric variable representing ...}
  \item{var5}{Numeric variable representing ...}
  \item{var6}{Numeric variable representing ...}
  \item{var7}{Numeric variable representing ...}
  \item{var8}{Numeric variable representing ...}
  \item{var9}{Numeric variable representing ...}
  \item{var10}{Numeric variable representing ...}
  \item{var11}{Numeric variable representing ...}
  \item{var12}{Numeric variable representing ...}
  \item{var13}{Numeric variable representing ...}
  \item{var14}{Numeric variable representing ...}
  \item{var15}{Numeric variable representing ...}
  \item{var16}{Numeric variable representing ...}
}
}
\source{
\url{http://example.com/data-source}
}
\usage{
ori_data
}
\description{
A comprehensive dataset used for various clustering and classification analyses.
}
\details{
The dataset includes multiple numeric features extracted from original observations.
}
\keyword{datasets}
