% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusterCenters.R
\name{plotClusterCenters}
\alias{plotClusterCenters}
\title{绘制聚类中心热图}
\usage{
plotClusterCenters(centers)
}
\arguments{
\item{centers}{聚类中心矩阵}
}
\value{
无返回值，直接绘制图形
}
\description{
展示聚类中心的特征值
}
\examples{
\dontrun{
plotClusterCenters(result$centers)
}
}
