% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotDecisionTreeStructure.R
\name{plotDecisionTreeStructure}
\alias{plotDecisionTreeStructure}
\title{绘制决策树结构图}
\usage{
plotDecisionTreeStructure(tree)
}
\arguments{
\item{tree}{决策树模型}
}
\value{
无返回值，直接绘制决策树结构图
}
\description{
使用 DiagrammeR 包绘制更美观的决策树结构
}
\examples{
\dontrun{
plotDecisionTreeStructure(tree)
}
}
