% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotKMeansResult.R
\name{plotKMeansResult}
\alias{plotKMeansResult}
\title{绘制 K-Means 聚类结果}
\usage{
plotKMeansResult(X, clusters)
}
\arguments{
\item{X}{数据矩阵，每行是一个样本}

\item{clusters}{聚类标签}
}
\value{
无返回值，直接绘制图形
}
\description{
使用 PCA 降维后对 K-Means 聚类结果进行可视化，并以红色叉形标记展示聚类中心
}
\examples{
\dontrun{
data(ori_data_x)
result <- kmeansR(ori_data_x, k = 3)
plotKMeansResult(ori_data_x, result$clusters)
}
}
