% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSpectralClusteringResult.R
\name{plotSpectralClusteringResult}
\alias{plotSpectralClusteringResult}
\title{绘制谱聚类结果}
\usage{
plotSpectralClusteringResult(X, clusters)
}
\arguments{
\item{X}{数据矩阵，每行是一个样本}

\item{clusters}{聚类标签}
}
\value{
无返回值，直接绘制图形
}
\description{
对谱聚类结果进行可视化展示
}
\examples{
\dontrun{
plotSpectralClusteringResult(ori_data_x, clusters)
}
}
